first attempt to make a workerThread with permanent data access moving files between paths correction of some error of last merge to this branch
