+++ src/main/resources/application.conf
@@ -1,7 +1,10 @@
 #raspberryGameSuiteServer properties
 server.port = 1500
 server.hostname = localhost
-user.properties.file.name = user.conf
+user.properties.file.name = user.sample.conf
+db.path = game_suite
+db.user = diogo
+db.pass = diogo
 
 #login client properties
 login.client.server.port = 1500
