--- src/main/resources/config.properties
@@ -1,7 +1,0 @@
-#raspberryGameSuiteServer properties
-port = 1500
-hostname = 127.0.0.1
-userPropsFileName = conf
-db.path = game_suite
-db.user = diogo
-db.pass = diogo
